feat : 뷰어 커스텀 스타일 적용
