+++ src/components/post/PostDetail/PostDetail.module.scss
@@ -1,0 +1,42 @@
+.post-detail {
+  .post-detail__title-group {
+    position: relative;
+    overflow: hidden;
+    color: white;
+    position: relative;
+    margin-top: 2rem;
+    height: 400px;
+    border-radius: 15px;
+
+    &:after {
+      content: '';
+      z-index: 0;
+      right: 0;
+      top: 0;
+      height: 100%;
+      width: 100%;
+      position: absolute;
+      background-color: rgba(0, 0, 0, 0.5);
+    }
+    .post-detail__title-box {
+      padding: 4rem 4rem 3rem;
+      height: 100%;
+      width: 100%;
+      display: flex;
+
+      flex-direction: column;
+      justify-content: space-between;
+      position: absolute;
+      z-index: 2;
+      .post-detail__title {
+        font-size: 2rem;
+      }
+      .post-detail__created-date {
+        text-align: end;
+        font-size: 1.2rem;
+      }
+    }
+  }
+  .post-detail__viewer {
+  }
+}
